refactor: replace `LocalComponentStore` with `LocalRouterComponentStore` implementation
