+++ packages/router-component-store/src/lib/local-router-store/provide-local-router-store.ts
@@ -1,6 +1,5 @@
 import { ClassProvider, Provider } from '@angular/core';
 import { RouterComponentStore } from '../router-component-store';
-import { LocalRouterComponentStore } from './local-router-component-store';
 import { LocalRouterStore } from './local-router-store';
 
 export function provideLocalRouterStore(): Provider {
@@ -9,5 +8,5 @@ export function provideLocalRouterStore(): Provider {
     useClass: LocalRouterStore,
   };
 
-  return [localRouterStoreProvider, LocalRouterComponentStore];
+  return localRouterStoreProvider;
 }
